+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Removed SUMMARY.md
 - Updtated composer.json to use phpunit 4.8
 - Downgraded dependency to 5.3
+- Added support for active / pasive connections in the FTP Adapter.
+- Added support for FTP_MODE_BINARY or FTP_MODE_ASCII in  the FTP Adapter (See examples/ftp.php) for details.
 
 
 ## 1.0  2015-12-27